rename processDiscoveredUrls to discoveredUrlProcessor
