+++ packages/web-api-scan-runner/src/crawl-runner/discovered-url-processor.spec.ts
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 import 'reflect-metadata';
 
-import { processDiscoveredUrls } from './process-discovered-urls';
+import { discoveredUrlProcessor } from './discovered-url-processor';
 
-describe(processDiscoveredUrls, () => {
+describe(discoveredUrlProcessor, () => {
     const urlsList = ['url1', 'url2', 'url3', 'url4'];
     const knownUrls = ['url1', 'url2', 'anotherUrl'];
     const maxUrlsLimit = 10;
@@ -12,7 +12,7 @@ describe(processDiscoveredUrls, () => {
     it('filters out known urls', () => {
         const expectedUrls = ['url3', 'url4'];
 
-        const processedUrls = processDiscoveredUrls(urlsList, maxUrlsLimit, knownUrls);
+        const processedUrls = discoveredUrlProcessor(urlsList, maxUrlsLimit, knownUrls);
 
         expect(processedUrls).toEqual(expectedUrls);
     });
@@ -20,7 +20,7 @@ describe(processDiscoveredUrls, () => {
     it('limits the number of urls according to config and count of knownUrls', () => {
         const urlCrawlLimit = 3;
 
-        const processedUrls = processDiscoveredUrls(urlsList, urlCrawlLimit, ['some url']);
+        const processedUrls = discoveredUrlProcessor(urlsList, urlCrawlLimit, ['some url']);
 
         expect(processedUrls.length).toBe(2);
     });
@@ -28,14 +28,14 @@ describe(processDiscoveredUrls, () => {
     it('filters and applies limit in correct order', () => {
         const urlCrawlLimit = knownUrls.length + 1;
 
-        const processedUrls = processDiscoveredUrls(urlsList, urlCrawlLimit, knownUrls);
+        const processedUrls = discoveredUrlProcessor(urlsList, urlCrawlLimit, knownUrls);
 
         expect(processedUrls.length).toBe(1);
         expect(knownUrls).not.toContain(processedUrls[0]);
     });
 
     it('handles missing knownPages', () => {
-        const processedUrls = processDiscoveredUrls(urlsList, maxUrlsLimit, undefined);
+        const processedUrls = discoveredUrlProcessor(urlsList, maxUrlsLimit, undefined);
 
         expect(processedUrls).toEqual(urlsList);
     });
@@ -43,7 +43,7 @@ describe(processDiscoveredUrls, () => {
     it('handles knownUrls.length > urlCrawlLimit', () => {
         const urlCrawlLimit = knownUrls.length - 1;
 
-        const processedUrls = processDiscoveredUrls(urlsList, urlCrawlLimit, knownUrls);
+        const processedUrls = discoveredUrlProcessor(urlsList, urlCrawlLimit, knownUrls);
 
         expect(processedUrls.length).toBe(0);
     });
